+++ hackathon_project/Assets/Scripts/Snake/ZoomScale.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 063700233f091438e8785898174c92f1
+guid: 9e6cf9b1212ae450780aa3ce1fbd862f
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
